Move Search Console utility functions to more appropriate location.
