+++ assets/js/modules/search-console/util/index.js
